requirements/local.txt
-r base.txt
django-debug-toolbar==4.2.0

# Linter
autoflake==2.2.0
black==23.3.0
isort==5.12.0
flake8==6.0.0
flake8-bugbear==23.6.5
flake8-comprehensions==3.13.0
flake8-eradicate==1.5.0
flake8-pie==0.16.0
pyflakes==3.0.1

